+++ tests/Fixtures/ShootResource/Pages/CreateShoot.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Blemli\Swissstreets\Tests\Fixtures\ShootResource\Pages;
+
+use Blemli\Swissstreets\Tests\Fixtures\ShootResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateShoot extends CreateRecord
+{
+    protected static string $resource = ShootResource::class;
+}