+++ tests/Fixtures/ShootResource/Pages/EditShoot.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Blemli\Swissstreets\Tests\Fixtures\ShootResource\Pages;
+
+use Blemli\Swissstreets\Tests\Fixtures\ShootResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditShoot extends EditRecord
+{
+    protected static string $resource = ShootResource::class;
+}