+++ tests/Fixtures/ShootResource/Pages/ListShoots.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Blemli\Swissstreets\Tests\Fixtures\ShootResource\Pages;
+
+use Blemli\Swissstreets\Tests\Fixtures\ShootResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListShoots extends ListRecords
+{
+    protected static string $resource = ShootResource::class;
+}